onoffweb/requirements.txt
Flask==3.0.2
icmplib==3.0.4
paramiko==3.4.0
PyYAML==6.0.1
wakeonlan==3.1.0
